fix: Cypress config again
Remove mountRouterComponent CT command and instead used MemoryRouter and RenderSearchParams directly in CT Credential test files to avoid any conflicts.
