+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "trainee-ui-app",
-  "version": "0.61.1",
+  "version": "0.61.2",
   "private": true,
   "dependencies": {
     "@aws-amplify/ui-react": "^4.2.1",
